+++ www/admin/controller/user/user.php
@@ -1,4 +1,5 @@
 <?php
+
 namespace Opencart\Admin\Controller\User;
 /**
  * Class User
@@ -352,6 +353,11 @@ class User extends \Opencart\System\Engine\Controller {
 
 		$data['sort'] = $sort;
 		$data['order'] = $order;
+
+		$this->document->addScript([
+			'view/javascript/oc/filter.min.js',
+			'view/javascript/oc/autocomplete.min.js'
+		]);
 
 		return $this->load->view('user/user_list', $data);
 	}