+++ www/admin/model/user/user.php
@@ -229,15 +229,15 @@ class User extends \Opencart\System\Engine\Model {
 		$implode = [];
 
 		if (!empty($data['filter_username'])) {
-			$implode[] = "LCASE(`u`.`username`) LIKE '" . $this->db->escape(oc_strtolower($data['filter_username']) . '%') . "'";
+			$implode[] = "LCASE(`u`.`username`) LIKE '%" . $this->db->escape(oc_strtolower($data['filter_username'])) . "%'";
 		}
 
 		if (!empty($data['filter_name'])) {
-			$implode[] = "LCASE(CONCAT(`u`.`firstname`, ' ', `u`.`lastname`)) LIKE '" . $this->db->escape('%' . oc_strtolower($data['filter_name']) . '%') . "'";
+			$implode[] = "LCASE(CONCAT(`u`.`firstname`, ' ', `u`.`lastname`)) LIKE '%" . $this->db->escape(oc_strtolower($data['filter_name'])) . "%'";
 		}
 
 		if (!empty($data['filter_email'])) {
-			$implode[] = "LCASE(`u`.`email`) LIKE '" . $this->db->escape(oc_strtolower($data['filter_email']) . '%') . "'";
+			$implode[] = "LCASE(`u`.`email`) LIKE '%" . $this->db->escape(oc_strtolower($data['filter_email'])) . "%'";
 		}
 
 		if (!empty($data['filter_user_group_id'])) {